fixup: remove windows-only tests from linux build
